requirements.txt
torch~=1.8.0
numpy~=1.19.2
blas~=1.0
tqdm~=4.50.2
torchvision~=0.9.0
distributed~=2.30.1
pandas~=1.1.3
pillow~=8.0.1
ignite~=0.4.4
matplotlib~=3.3.2
scikit-learn~=0.23.2
ipython~=7.19.0